Redesign command buffers around submission-time IR
Replace command contexts with queue-independent command buffers that retain Rutile IR until submission. Virtualize queues, encode stable opaque timepoints, and lower recorded waits into ordered native submission segments with retained resources and explicit synchronization.

Bring Vulkan, DirectX 12, and OpenGL onto the same command model, including vertex streams, descriptor and resource bindings, state/layout transitions, validation and logging dispatch, loader and extension integration, and runtime deployment for all example backends.

Consolidate the public binding layout under include/ and source/, removing obsolete binding trees, stale command-context files, compatibility wrappers, and unused OpenGL GLSL paths. OpenGL now requires GL 4.6 core and uses the SPIR-V path only.

+++ runtime/layers/rt-validation-layer/src/buffer.h
@@ -9,8 +9,8 @@ struct rtval_buffer {
 
 struct rtval_buffer* rtval_buffer_create(void);
 void rtval_buffer_destroy(struct rtval_buffer* buffer);
-rt_timepoint rtval_buffer_data(struct rtval_buffer* buffer, enum rt_buffer_mode mode, enum rt_buffer_usage usage, u64 size, const void* data);
-rt_timepoint rtval_buffer_subdata(struct rtval_buffer* buffer, u64 offset, u64 size, const void* data);
+rt_timepoint rtval_buffer_data(struct rtval_buffer* buffer, enum rt_buffer_mode mode, enum rt_buffer_usage usage, usize size, const void* data);
+rt_timepoint rtval_buffer_subdata(struct rtval_buffer* buffer, usize offset, usize size, const void* data);
 void rtval_buffer_read(struct rtval_buffer* buffer, u64 offset, u64 size, void* data);
 
 #endif
